requirements.txt
selenium~=3.141.0
tqdm~=4.62.3
SQLAlchemy~=1.4.25
pandas~=1.3.3
numpy~=1.21.2
requests~=2.26.0
openpyxl~=3.0.9
redis~=3.5.3
concurrent_log_handler~=0.9.19
elasticsearch~=7.15.0
elasticsearch_dsl~=7.4.0
kafka-python~=2.0.2
